+++ src/main/java/woojooin/planitbatch/batch/job/TempJob.java
@@ -1,0 +1,4 @@
+package woojooin.planitbatch.batch.job;
+
+public class TempJob {
+}
