+++ src/main/java/woojooin/planitbatch/batch/processor/TempProcessor.java
@@ -1,0 +1,4 @@
+package woojooin.planitbatch.batch.processor;
+
+public class TempProcessor {
+}
