+++ src/main/java/woojooin/planitbatch/batch/reader/TempReader.java
@@ -1,0 +1,4 @@
+package woojooin.planitbatch.batch.reader;
+
+public class TempReader {
+}
