+++ src/main/java/woojooin/planitbatch/batch/scheduler/JobScheduler.java
@@ -1,8 +1,6 @@
-package woojooin.planitbatch.scheduler;
+package woojooin.planitbatch.batch.scheduler;
 
 import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobParameters;
-import org.springframework.batch.core.JobParametersBuilder;
 import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
