+++ src/main/java/woojooin/planitbatch/batch/scheduler/TempScheduler.java
@@ -1,0 +1,4 @@
+package woojooin.planitbatch.batch.scheduler;
+
+public class TempScheduler {
+}
