+++ src/main/java/woojooin/planitbatch/batch/writer/TempWriter.java
@@ -1,0 +1,4 @@
+package woojooin.planitbatch.batch.writer;
+
+public class TempWriter {
+}
