--- src/main/java/woojooin/planitbatch/config/TempConfig.java
@@ -1,4 +1,0 @@
-package woojooin.planitbatch.config;
-
-public class TempConfig {
-}
