--- src/main/java/woojooin/planitbatch/job/TempJob.java
@@ -1,4 +1,0 @@
-package woojooin.planitbatch.job;
-
-public class TempJob {
-}
