--- src/main/java/woojooin/planitbatch/processor/TempProcessor.java
@@ -1,4 +1,0 @@
-package woojooin.planitbatch.processor;
-
-public class TempProcessor {
-}
