--- src/main/java/woojooin/planitbatch/reader/TempReader.java
@@ -1,4 +1,0 @@
-package woojooin.planitbatch.reader;
-
-public class TempReader {
-}
