--- src/main/java/woojooin/planitbatch/scheduler/TempScheduler.java
@@ -1,4 +1,0 @@
-package woojooin.planitbatch.scheduler;
-
-public class TempScheduler {
-}
