--- src/main/java/woojooin/planitbatch/writer/TempWriter.java
@@ -1,4 +1,0 @@
-package woojooin.planitbatch.writer;
-
-public class TempWriter {
-}
